repository: Replicate object pools during `ReplicateRepository` RPC
When replicating a repository to a different storage, if the source
repository is linked to an object pool, the object pool relationship
should also be recreated on the target storage. With ef1a7956, object
pool links are replicated on the target storage if the object pool
already exists on the target storage.

Update object pool syncing in the `ReplicateRepository` RPC to create
the object pool from a snapshot of the source object pool if it does not
exist on the target storage.