gitaly/diff: Fix parsing of type-changes with SHA256
In de9712689 (gitaly/diff: Allow parsing of diffs generated with SHA256
object format, 2023-07-20), we have introduced support for parsing diffs
which are generated in repositories that use the SHA256 object format.
One oversight though is that we still fail to parse diffs that have a
type change, e.g. from a regular file to a symlink, because we are in
fact lacking test coverage for parsing such diffs.

Fix this by using the correct object format when handling type changes
and add a testcase to verify that it works as expected.