diff: Ready CommitDiff tests for the SHA256 object format
Ready the CommitDiff tests for the SHA256 object format by generating
test data at runtime instead of using pre-seeded repositories.

Note that this refactoring boils down to an almost-rewrite for the limit
tests. The previous version of those tests was almost impossible to
understand. The newer version is much simpler now and should hopefully
address that shortcoming.