ci: Add variable to force pipeline to run
For debugging CI it is often useful to be able to run a pipeline outside
of the context of an MR. The GitLab project has a `FORCE_GITLAB_CI`
variable that forces a pipeline to run even when the workflow
requirements are not met.

Add an equivalent `FORCE_GITALY_CI` variable to allow us to manually
run pipelines as needed.