gittest: Expected exit code for command
There are sub-commands in git that may return non-zero exit code.
One of them is diff. It compares files and returns comparison result.
If there is a difference found the exit code won't be 0. In some
cases it is a legit and all we want is to check the output. So now
there is an ExpectedExitCode field that can be used to set expected
exit code for the command. If exit code doesn't match the test will fail.