ci: Drop optimization to run tests with code changes only
We have recently started an experiment where we changed our tests jobs
to only run in case there are some actual code changes. This change has
proven to be very fragile as we had many contexts now where we didn't
run any test jobs despite some code changes.

The original intent of this change was to skip running tests when we
know that there were only documentation changes. Our CI rules don't have
such a condition though, which means that we are forced to express the
opposite rule that there were only code changes. This has the downside
though that we now fail open (that is we err on the side of running too
few CI jobs) instead of failing closed (which would be to rather run too
many CI jobs). This is not a good failure mode though as it is now quite
easy for us to miss test failures introduced by a merge request.

Drop the optimization again and always run tests, regardless of which
files have changed.