operations: Simplify error checking in UserCommitFiles tests
The error checking in UserCommitFiles is hard to follow as a subset of
our tests allow passing multiple different expected errors. This has two
reasons:

    - We still assert an `indexError`, which is a legacy error type that
      existed in the past when errors were still embedded in successful
      responses. This has since been dropped, but we didn't clean up the
      now-unused error.

    - Many of the tests return a structured error with metadata, which
      is a bit more involved to set up. The tests thus assemble the
      error for the testcase when a `structError` is set.

The first is trivial to simplify by just dropping the unused error. The
latter is addressed by introducing `structuredIndexError()`, which is a
helper function that assembles the structured error.