Generate transaction's pack file directly in correct location
TransactionManager computes a pack file that includes all objects
from the quarantine that are needed by the new reference tips. This
generated pack file may end up with the same name as one of the packs
already present in the quarantine if the content matches exactly the
content of a pack in the quarantine directory. This would generally be
fine, as Git names the pack files based on their content. If they have
they have the same name, they should have the same content. However,
when the transaction is staged, the generated pack file is moved from
the quarantine to the 'wal-files' directory that gets ultimately logged.
If the generated pack file's name matched a quarantined pack's name, the
quarantine is left without any pack file after the move. This then fails
the reference verification checks as the new objects can't be found in
the quarantine post move.

This commit fixes the issue by writing the generated pack directly into
the 'wal-files' directory which ultimately gets logged. This way the
quarantine and the generated packs stay free of any sort of conflicts.
Things also become simpler as we don't separately have to move the files
into place.

The pack is staged using a pre-defined name 'objects.pack'. The existing
tests are updated to remove the hash dependent assertions given the name
of the logged pack is now constant. The pack is still applied into the
Git repository with the Git generated name to deduplicate packs that are
exactly the same.