localrepo: Delete parent entry when last child is deleted
If there is a tree entry with a single child entry, deleting that child
doesn't delete the parent entry. This is the expected behavior from the
initial design perspective, since the parent would be allocated an empty
tree ID.

But this breaks our RPCs as generally deleting the last file of a folder
should also delete the folder. There's been extensive discussion around
this topic on the mailing list in the past, and the agreed-upon result
was always that Git will not add empty directories. Keeping in line with
this, we should delete parent tree entries when its last child has been
deleted.