operations: Split up `merge.go` per contained RPC
The `merge.go` file contains the implementations for UserMergeBranch,
UserMergeToRef as well as UserFFBranch. This results in a rather long
file with semi-related functionality, makes the implementations of the
respective RPCs hard to find and makes it hard to convert their tests to
support SHA256 one by one.

Split up the file into per-RPC code units as well as a `merge_utils.go`
file that contains common functionality required for merges.