limiter: Use CountTicker instead of ManualTicker in calculator tests
This test was flaky because the calibration goroutine started later than
expected. When `Start()` function exits, the test starts to tick the
manual timer. If the calibration goroutine starts too late, the
calibration may miss one tick. Hence, this commit converts the test to
use CountTicker, which allows a callback after the timer is reset for N
times. It also adds a simple synchronization channel to ensure `Start()`
returns when the calibration goroutine already starts.