Makefile: Remove cleanup of ruby/
We deleted a bunch of files and directories related to the Ruby sidecar,
but some of them were mentioned in .gitignore. This caused git to never
remove them. For that reason we introduced a target `clean-ruby` which
was called by `build` and `clean` to make sure these files get deleted
some way or the other.

Because that target has been around for a couple of releases now, it
should have done it's duty, and it's safe to retire it.

Milestone: 16.3
Label: maintenance::removal