localrepo: Introduce `localrepo.CloneBundle` method
When creating a repository from a bundle, it is necessary to detect the
object format of the repository to ensure the correct repository is
initialized. To facilitate this, the `localrepo.CloneBundle` method is
introduced which performs a `git-clone(1)` on a Git bundle, thus
automatically initializing the repository with the same object format
used by the bundle.