repository: Make `CreateRepositoryFromBundle` RPC clone from bundle
Currently the `CreateRepositoryFromBundle` RPC initalizes a new Git
repository and performs `git-fetch(1)` on a Git bundle to seed the
repository. If the Git bundle object format is not SHA1, the bundle will
fail to be extracted because the created Git repository always uses the
default SHA1 object format.

Skip initial Git repository creation and perform `git-clone(1)` on the
bundle to delegate repository initialization to Git so correct object
format is selected automatically.