testserver: Redirect Praefect's output to a log file
When executing the `test-with-praefect` test target, then we execute
the Praefect binary as a separate process and have it proxy all requests
to Gitaly. While the standard error of this process is redirected to a
buffer, its standard output uses the standard output used by our tests.
This causes us to write tons of log messages to stdout, which makes the
overall output almost unreadable. Furthermore, this causes our CI to
choke because we bust the limit of how many bytes of log output may be
generated.

Follow the same approach as we use for the Gitaly test server and
redirect its standard output to a log file instead.