Unify closing method and field names
PartitionManager and TransactionManager are using deviating names to
signal scenarios when the closing process has been initiated and when
it has completed. This commit unifies the names in both types and moves
to use `close` for the method to initiate closing, `closing` for the
channel that signifies the process has been initiated, and `closed` for
the final end state when the closing has finished.