operations: Inline UserUpdateSubmodule logic
Not long ago we have converted UserUpdateSubmodule to use plain Git
instead of libgit2. This was implemented via two separate funcitons that
were called depending on whether the corresponding feature flag was
enabled or disabled. Nowadays we only have the Git implementation left,
but have retained the indirection via a separate `userUpdateSubmodule()`
function.

Let's inline this function to simplify the code a small bit. This will
also make it easier to detect the object hash early on and pass it into
the request validation logic.