counter: Add RepositoryCounter
Gitaly currently does not expose a way to determine the number of
repositories present on a storage. In a cluster an admin could query the
Praefect database, but this is undocumented and is not available for
non-cluster nodes.

Introduce a `RepositoryCounter` to manage the new
`gitaly_total_repositories_count` gauge. For our purposes object pools
are considered to be separate repositories and no effort is made to
associate them with their children.

The metric will include the storage name and the first relevant
directory of the repository path as `prefix`. The latter is included to
allow admins to understand which types of repositories are present. For
a standalone gitaly node `prefix` would be the first directory up
from the storage, e.g. `@hashed`, `@pools`, or `@snippets`. For a Gitaly
Cluster node it would be the child directory of `@cluster`, e.g.
`repositories` or `pools`.