counter: Wire up RepositoryCounter to repoutil
All repository creation and removal is done via `repoutil`, so
hooking `RepositoryCounter` in here will allow us to track changes
without dealing with the idiosyncracies of the many RPCs that may
create repositories.

Add a `RepositoryCounter` as a parameter to the `repoutil.Create` and
`repoutil.Remove` functions and wire it up to necessary services.

The exception to this is the `RemoveAll` RPC which deletes an entire
storage, we use the counter directly here.