praefect: Wire `praefect.Coordinator` to `NewGRPCServer()`
In a future commit, a Praefect stream handler will be registered for the
`ReplicateRepository` RPC. This handler will need to use
`praefect.Coordinator` for coordinating RPC transactions and routing.

Make the necessary adjustments to wire `praefect.Coordinator` to
`NewGRPCServer()` where the new stream handler will be registered in the
future.