praefect: Introduce `ReplicateRepositoryHandler()`
In order to support object pool replication in Praefect, the
`ReplicateRepository` RPC needs to be intercepted and utilize a custom
stream handler. To facilitate this, the `ReplicateRepositoryHandler()`
function is introduced. At this point, the handler redirects back to the
default `proxy.TransparentHandler()`. In future commits, the handler
will be extended to also support object pool replication.