repository service: Change BackupRepository tests to check for refs
These RPC tests are only smoke testing that the backups get created
properly. For this RPC, if there are no errors, then a refs file will
always exist, but a bundle may not exist. So for the sake of ease, we
swap the final check to look for refs instead of a bundle.