catfile: Recognize SHA256 GPG signatures
Commit signatures are usually denoted by a "gpgsig" field in the commit.
When using the SHA256 object format though, Git uses "gpgsig-sha256" as
field name instead. We do not recognize this field name in the catfile
package yet and will thus fail to parse such commits.

Add the logic to detect this field as well as a test to verify that we
can parse such commits as expected.