commit: Fix parsing of gpgsig fields with trailing data
In order to extract commit signatures we scan the commit message for
"gpgsig" fields. Starting from any line with such a prefix we then scan
until we find a line that terminates the signature, e.g. an empty line.

The logic to do this is buggy though as we didn't include the trailing
space that separates the field name from its data. Consequentially, we
currently recognize all fields that start with "gpgsig" as signature
data, even if they have an arbitrary suffix like "-garbage". This suffix
is instead appended to the signed commit data that we extract.

Fix this bug by checking for the space. While at it, prepare the code
for SHA256 support already by reformatting it a bit. Signatures in
SHA256 repositories use the "gpgsig-sha256" field instead, which is how
this bug was indeed found.

Changelog: fixed