config: Add JSON tags to configuration members
We're about to add support for a configuration command to Praefect that
allows the user to compute parts of the configuration via an external
executable. As TOML is not quite the best fit to programmatically carry
data between processes, we'll instead be using JSON. Many of the members
part of the configuration structures do not yet have JSON tags though.

Add JSON tags so that the respective names for each of the members are
well-defined.