Implement a transaction registry
While generally the transaction would be accessible directly within a
request's context, transactions need to be referenceable across the process
boundary as well. Hooks for example involve stepping out from the main
Gitaly process and back in. When executing the hooks, we need to be able
to retrieve the transaction this hooks call is connected to. This commit
implements a simple registry where the transactions can be stored when
they are began and retrieved from later on. The hook payload can carry
the transaction ID and use it to retrieve the transaction object from
the registry when executing the hook.

`TransactionID` type is in the storage package to avoid cyclic
dependencies between the packages that will need it.