Methods for storing transaction ID in context
Transaction's ID need to be piped through to the hooks. In order to pipe
the transaction ID through the various functions on the call path, this
commit implements helpers to include the transaction ID in the context.
This way it can be piped all the way to the location for inclusion in
the hook payloads.