Capture reference changes in transaction through the hooks
Transactions will write their changes into their snapshots. To commit
those changes into the actual repository, the reference changes done
in the snapshot need to be captured for committing as part of the
transaction. This commit captures the reference changes made through
the reference transaction hook.

The references are captured in two phases. In the `prepare` phase,
only the old value of the reference captured. In the `committed` phase,
the actual reference updates have been performed and are thus are staged.

The old value of the reference is captured in the `prepare` phase as
we wouldn't be able to determine the old value anymore in the `commited`
phase. The reference transaction hook is invoked with the zero OID as
the old OID for forced reference updates which don't care about the old
value of the reference. Once the update is done, we wouldn't be able to
read the old value anymore in the `committed` phase.

Forced reference updates are converted to normal reference updates and
the old OID is resolved to do so. This is done as there's ambiguity
between forced create-or-update reference and normal reference creations
as they both would be represented with zero OID in the reference
transaction hook. This still allows callers to do reference updates
without knowing the current reference value but forbids racy updates
where concurrent transactions changes the reference value. This should be
fine given racy updates would lead to undeterminable outcomes anyway.