signature: Move signature to internal/git
The type `Signature` is originally used by git2go to help generating
signatures for committer and author of a commit. The package git2go
would reach the end of its life so we should put `Signature` to a more
suitable package for future use.

And the declaration of `Signature` and `NewSignature` in git2go are kept
to avoid too many code changes.

Signed-off-by: Xing Xin <xingxin.xx@bytedance.com>