Don't close a partition twice
PartitionManager currently has an issue where it can panic if a partition
is closed twice. This may happen if the partition's last transaction
finishes at the same time as the PartitionManager itself is being closed.
While there's no data race due to the closing being done under a lock, the
closing may be done twice. Fix the issue by checking whether the partition
has already been closed or not before closing the channel.