localrepo: Add a basic rebase support
This commit provides a rudimentary rebase functionality implemented in
pure git, which can be used as a drop-in replacement for git2go in
UserRebaeToRef and UserRebaseConfirmable.

When Gitaly performs a rebase, it applies the commits from a remote
branch onto a local branch, excluding commits already existing on the
local branch [1]. The commits are kept in original order [2]. Therefore,
to implement a basic rebase, we need to:

1. Find commits that should be rebased, preserving their order.
2. Apply each commit one-by-one using cherry-pick.

The `do not skip empty commit` branch of `TestRebase_skipEmptyCommit`
from git2go [3] is omitted because currently we are not using that
option anywhere.

[1] Merge commits and already cherry-picked commits are excluded.
[2] Specifically a topo order, see https://github.com/git/git/blob/v2.41.0/sequencer.c#L5712
[3] Refer to tests in cmd/gitaly-git2go/rebase_test.go

Inspired-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>
Inspired-by: Jiuyang Xie <jiuyang.xjy@alibaba-inc.com>
Helped-by: Han Xin <hanxin.hx@bytedance.com>
Signed-off-by: Xing Xin <xingxin.xx@bytedance.com>