Configure transaction middleware in test server
This commit configures the transaction middleware in Gitaly's
test server when GITALY_TEST_WAL environment variable is set. With
the middleware in place, tests that are using the test server are
automaticaclly ran with the transaction management enabled.

Mutators which perform changes that are not automatically recorded
in the transaction are updated to record the changes to have the tests
pass. This includes RemoveRepository to record the deletion, UserSquash
to record the unreference commit object, SetCustomHooks to record the new
hooks and default branch updates.

Some tests were adapted to the changed behavior with transactions. The
behavior changes into the following categories:

1. Lock conflicts don't occur anymore with the transactions running in
   isolated snapshots. Such tests were disabled if transactions are
   enabled.
2. Concurrent reference update errors no longer manifest during the RPC
   handler's execution as the handlers are isolated from concurrently
   performed changes. Tests testing this behavior were updated to expect
   the reference conflict error from the transaction commit.
3. Objects are not written into the repository if the transaction doesn't
   commit. Tests which are asserting objects were migrated after the
   'pre-receive' hook succeeded have been updated to assume that no
   objects make it into the repository if the transaction doesn't commit.
4. The handlers receive a rewritten repository that points to the snapshot.
   Tests which are asserting the relative path or the quarantine variables
   have been updated to not do so.

Tests which are utilizing object pools have been disabled as object pools
are not yet supported by the transaction manager.