limiter: Integrate adaptive limit to pack-objects limiter
This commit integrates the adaptive limit to pack-objects limiter of the
Gitaly server. It introduces some new configurations for pack-objects
limiting. When Gitaly server starts, it creates an adaptive calculator.
The calculator adjusts the adaptive limit which is already integrated
into the concurrency limit in prior commits.

The integration is hidden behind `gitaly_adaptive_limit_pack_objects`
feature flag. If this flag is not enabled, the limiter uses a static
limit from `max_concurrency` config.