Makefile: test-gitaly-linters requires gotestsum
The `test-gitaly-linters` target requires `gotestsum` to be present to
execute successfully, but does not list it as a dependency. In typical
operation in CI this is not an issue as we always run tests first, but
running it standalone on a clean worktree will fail.

Add `${GOTESTSUM}` as a dependency for the task.