backup: Always skip missing backups on restore
In practice this option never worked because when we go to backup a
repository that doesn't exist, then nothing gets written to the backup.
Then on restore, since there was no backup, the fallback locator was
used, and then the missing backup was skipped.

Since there is no other option other than skipping when a backup is not
found, we can remove this option.

The tests are not exactly accurate to production as they are not using a
fallback. Hence there are some changed tests.