operations: Don't use hardcoded object IDs in UserDeleteBranch tests
One of the UserDeleteBranch tests is still using hardcoded object IDs,
which is going to be a problem when we want to test with SHA256. While
we could just use `gittest.ObjectHashDependent()` and hardcode both SHA1
and SHA256 object IDs, this would be a test smell.

While we could just put the complete service instantiation into a setup
function, this would be quite verbose. Instead, we only parameterize the
`expectedErr` function so that it gets the expected old object ID as
input. While not exactly pretty, it at least works.